+++ packages/micromanage/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.2.5](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/micromanage-cli@1.2.4...@carbon-platform/micromanage-cli@1.2.5) (2022-03-01)
+
+
+### Bug Fixes
+
+* add schema readme ([5381484](https://github.com/carbon-design-system/carbon-platform/commit/53814840e33955ec6eb5cff3d3d1f72eac23b01f))
+
+### [1.2.4](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/micromanage-cli@1.2.3...@carbon-platform/micromanage-cli@1.2.4) (2022-03-01)
+
 ### [1.2.3](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/micromanage-cli@1.2.2...@carbon-platform/micromanage-cli@1.2.3) (2022-03-01)
 
 ### [1.2.2](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/micromanage-cli@1.2.1...@carbon-platform/micromanage-cli@1.2.2) (2022-03-01)