+++ services/web-app/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.67](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.66...@carbon-platform/web-app@0.1.67) (2022-03-01)
+
+### [0.1.66](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.65...@carbon-platform/web-app@0.1.66) (2022-03-01)
+
+### [0.1.65](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.64...@carbon-platform/web-app@0.1.65) (2022-03-01)
+
+### [0.1.64](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.63...@carbon-platform/web-app@0.1.64) (2022-03-01)
+
+
+### Features
+
+* add illo to mobile hero component ([15da76c](https://github.com/carbon-design-system/carbon-platform/commit/15da76c3202787fabbe2f276cff927365417b57f))
+
+### [0.1.63](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.62...@carbon-platform/web-app@0.1.63) (2022-03-01)
+
+### [0.1.62](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.61...@carbon-platform/web-app@0.1.62) (2022-03-01)
+
 ### [0.1.61](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.60...@carbon-platform/web-app@0.1.61) (2022-03-01)
 
 ### [0.1.60](https://github.com/carbon-design-system/carbon-platform/compare/@carbon-platform/web-app@0.1.59...@carbon-platform/web-app@0.1.60) (2022-03-01)